Move ViewModels to correct package
