+++ noty-android/app/src/main/java/dev/shreyaspatil/noty/view/viewmodel/AboutViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.shreyaspatil.noty.simpleapp.view.viewmodel
+package dev.shreyaspatil.noty.view.viewmodel
 
 import androidx.lifecycle.ViewModel
