+++ noty-android/app/src/main/java/dev/shreyaspatil/noty/view/viewmodel/HomeViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.shreyaspatil.noty.simpleapp.view.viewmodel
+package dev.shreyaspatil.noty.view.viewmodel
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.ViewModel
