+++ noty-android/viewmodel/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="dev.shreyaspatil.noty.simpleapp.view.viewmodel" />
+<manifest package="dev.shreyaspatil.noty.view.viewmodel" />
